Add Node.get_grandparent()
With documentation and unit testing.

Node.xml should be alphabetical because doctool doesn't like it if the docs aren't alphabetically sorted.

New commits should either keep this level of quality or exist on a different branch.

Also only commits that make sense for the branch should exist such as small complete features. More in README.md.