Hacky CodeEdit fix for #95872
Note: I'm unsure if one of code_completion_prefixes being empty would cause issues. However i have added a config warning.

Also i have replaced empty keys with # or " " in _set_delimiters as to prevent the empty errors. However i highly suspect testcase failures.

This is definitely not a correct fix as the non-empty set is a hack just for the editor because the bug does not appear in code, but the fix alters how code works, requiring a rewrite of the tests.